Update the like button and test as well as refactor some codes
